Add files via upload
These two files should allow to plot CUE for individual taxa. Specific changes include: - grid.py: i) activation of lines 403-408, ii) activation of line 93 where the variable CUE_taxon is defined. It might not be needed because without prior definition, it is still stored and can be called during the simulations. -output.py: i)activation of lines 246-249 to create and store the variable CUE_TaxaSeries. I side note, this file has some values over 1 but this is just because some taxa have more than 1 count. I cannot divide this in the output.py file, therefore, I stored an extra variable called Taxon_counts that has the counts. The final calculation of the CUE is done later in the postprocesing.